# admin-ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "arctern",
    "description": "ZFS replication daemon HTTP API",
    "license": {
      "name": ""
    },
    "version": "0.3.0"
  },
  "paths": {
    "/api/v1/config": {
      "get": {
        "tags": [
          "config"
        ],
        "operationId": "get_config",
        "responses": {
          "200": {
            "description": "The TOML file currently loaded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfigView"
                }
              }
            }
          },
          "500": {
            "description": "Config file unreadable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets": {
      "get": {
        "tags": [
          "datasets"
        ],
        "summary": "List datasets through the daemon's shared typed ZFS facade. It uses\nRealRunner in production and the SSH test runner only for dev/integration.",
        "operationId": "list_datasets",
        "responses": {
          "200": {
            "description": "All datasets visible to the daemon's ZFS runner",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetSummary"
                  }
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{name}/holds": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "`zfs holds -p -H <snap>...` over every snapshot of `{name}` in one\n(chunked) invocation. The per-snapshot endpoint costs one `zfs holds`\nper row, so a browser showing a dataset with hundreds of snapshots\nturns every refresh into hundreds of process spawns — through the\ncontrol channel when the dataset lives on a peer. This is the shape\nthe snapshot browser uses.",
        "description": "Snapshots without holds are absent from the map. The response always\ncovers the whole dataset, so a missing key means \"no holds\", never\n\"not fetched\".",
        "operationId": "list_dataset_holds",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshot tag to its holds, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/SnapshotHold"
                    }
                  },
                  "propertyNames": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{name}/snapshots": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "List snapshots of `{name}` (one dataset, non-recursive). Properties\n`creation` and `used` come along so the UI can render age + size.",
        "operationId": "list_snapshots",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "prefix",
            "in": "query",
            "description": "Filter to snapshots whose tag (after the `@`) starts with this\nprefix. Useful to keep zrepl_-style snapshots out of the noise\nwhen manual snapshots also exist on the dataset.",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Snapshots of the dataset, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DatasetSummary"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Dataset not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Create a snapshot of `{name}` named `req.snapshot_name`.\n`zfskit::ZfsError::SnapshotExists` maps to 409 — the caller\ndecides whether already-exists is fatal.",
        "operationId": "create_snapshot",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSnapshotRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Snapshot created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatasetSummary"
                }
              }
            }
          },
          "404": {
            "description": "Parent dataset not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Snapshot already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{name}/snapshots/{snapshot}/destroy": {
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Destroy snapshot `{name}@{snapshot}`. Path-segment escaping: the\ndataset goes URL-encoded (`%2F` for `/`); the snapshot tag is the\npart after the `@`.",
        "operationId": "destroy_snapshot",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "description": "Snapshot tag (the part after `@`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Snapshot destroyed"
          },
          "404": {
            "description": "Snapshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "409": {
            "description": "Snapshot has a hold",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/{name}/snapshots/{snapshot}/holds": {
      "get": {
        "tags": [
          "snapshots"
        ],
        "summary": "`zfs holds <snapshot>` — list user holds blocking destroy. Empty\nvec means the snapshot is destroy-eligible (as far as holds are\nconcerned; sub-clones still block independently).",
        "operationId": "list_holds",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "description": "Snapshot tag (the part after `@`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Holds on the snapshot, oldest first",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SnapshotHold"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Snapshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "snapshots"
        ],
        "summary": "Place a user hold on `{name}@{snapshot}`. Tags with the `arctern_`\nprefix are refused — they'd collide with the replication machinery's\nstep/last holds and be swept or misread by it.",
        "operationId": "create_hold",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Parent dataset (URL-encode `/` as %2F)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "snapshot",
            "in": "path",
            "description": "Snapshot tag (the part after `@`)",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateHoldRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Hold placed"
          },
          "400": {
            "description": "Invalid or reserved tag",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "Snapshot not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "ZFS returned an error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Subscribe to the daemon's log-event broadcast and yield each as an\nSSE frame, preceded by a replay of the most recent events so a\nfreshly opened page shows context instead of an empty feed until\nsomething new happens. A reconnecting client that names its cursor\ngets only what it missed.",
        "operationId": "stream_events",
        "parameters": [
          {
            "name": "since",
            "in": "query",
            "description": "Resume after this event id. A client that already holds part of\nthe log passes its newest id so the replay is not re-delivered.",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream of LogEvent JSON frames"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiErrorBody": {
        "type": "object",
        "description": "Body shape for `4xx`/`5xx` responses from the daemon. `error` is a\nshort machine-readable category (`spawn`, `dataset_not_found`, …);\n`message` is a human-readable description.",
        "required": [
          "error",
          "message"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ConfigView": {
        "type": "object",
        "description": "Body of `GET /api/v1/config` — the on-disk TOML the daemon was\nstarted with, plus its absolute path. Read-only.",
        "required": [
          "path",
          "content_toml"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "content_toml": {
            "type": "string"
          }
        }
      },
      "CreateHoldRequest": {
        "type": "object",
        "description": "Body of `POST /api/v1/datasets/{name}/snapshots/{snapshot}/holds`.\n`arctern_*` tags are reserved for the replication machinery and\nrejected.",
        "required": [
          "tag"
        ],
        "properties": {
          "tag": {
            "type": "string"
          }
        }
      },
...
      "DatasetSummary": {
        "type": "object",
        "description": "Slim projection of a `zfs list` entry suitable for HTTP + OpenAPI.\nNative ZFS properties carry typed data (bytes, bool, …) but\n`BTreeMap<String, String>` serializes more cleanly through utoipa;\nconsumers parse property values as needed.",
        "required": [
          "name",
          "dataset_type"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "dataset_type": {
            "$ref": "#/components/schemas/DatasetType"
          },
          "properties": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            },
            "propertyNames": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}